descriptors: move the LianaDescKey to the keys submodules
